+++ app/Http/Controllers/CommentController.php
@@ -4,12 +4,16 @@ namespace App\Http\Controllers;
 
 use App\Comment;
 use App\Http\Helpers\PermissionHandler;
+use App\Http\Helpers\Validator;
+use App\Http\Traits\SearchTrait;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\URL;
 
 class CommentController extends Controller
 {
+    use SearchTrait;
+
     public function __construct()
     {
         $this->middleware('auth')->except('show');
@@ -23,19 +27,10 @@ class CommentController extends Controller
     public function index()
     {
         //
-        PermissionHandler::noCommentEditorAbort();
-        $comments = Comment::all();
-        return view('comment.index', compact('comments'));
-    }
+        $this->authorize('viewAny', Comment::class);
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
+        $comments = Comment::paginate(50);
+        return view('comment.index', compact('comments'));
     }
 
     /**
@@ -47,6 +42,8 @@ class CommentController extends Controller
     public function store(Request $request)
     {
         //
+        Validator::validate($request, 'comment');
+
         $comment = new Comment();
         $comment->article_id = $request->article;
         $comment->user_id = Auth::id();
@@ -82,10 +79,7 @@ class CommentController extends Controller
     public function edit(Comment $comment)
     {
         //
-        if ($comment->user_id !== Auth::id())
-        {
-            PermissionHandler::notEditCommentAbort();
-        }
+        $this->authorize('update', $comment);
         return view('comment.edit', compact('comment'));
     }
 
@@ -99,10 +93,8 @@ class CommentController extends Controller
     public function update(Request $request, Comment $comment)
     {
         //
-        if ($comment->user_id !== Auth::id())
-        {
-            PermissionHandler::notEditCommentAbort();
-        }
+        $this->authorize('update', $comment);
+        Validator::validate($request, 'comment_update');
 
         $comment->comment = $request->comment;
 
@@ -120,6 +112,7 @@ class CommentController extends Controller
     public function destroy(Comment $comment)
     {
         // Should be avoided, as it causes cascade delete on child comments.
+        $this->authorize('delete', Comment::class);
         PermissionHandler::notDeleteCommentAbort();
         Comment::where('id', $comment->id)->delete();
 
@@ -131,21 +124,41 @@ class CommentController extends Controller
     }
 
    /**
-    * @param  \Illuminate\Http\Request  $request
     * @param  \App\Comment  $comment
     * @return \Illuminate\Http\Response
     */
-    public function delete(Request $request, Comment $comment)
+    public function delete(Comment $comment)
     {
         // Default way to delete comments, this is a non destructive of deleting comments in order to preserve child comments.
-        if ($comment->user_id !== Auth::id())
-        {
-            PermissionHandler::notEditCommentAbort();
-        }
-        $comment->comment = "Comment has been deleted.";
+        $this->authorize('update', $comment);
 
+        $comment->comment = "Comment has been deleted.";
         $comment->save();
 
         return redirect()->back();
+    }
+
+    public function searchComments(Request $request)
+    {
+        $this->authorize('viewAny', Comment::class);
+
+        $columns = ['id', 'article_id', 'user_id', 'parent_id', 'comment', 'created_at', 'updated_at'];
+        $query = Comment::select();
+        $comments = $this->search($query, $columns, $request->keyword, true, 50);
+        return view('comment.index', compact('comments'));
+    }
+
+    public function userComments()
+    {
+        $comments = Comment::where('user_id', Auth::id())->paginate(50);
+        return view('comment.index', compact('comments'));
+    }
+
+    public function searchUserComments(Request $request)
+    {
+        $columns = ['id', 'article_id', 'parent_id', 'comment', 'created_at', 'updated_at'];
+        $query = Comment::select();
+        $comments = $this->idRestrictedSearch($query, $columns, $request->keyword, true, 50);
+        return view('comment.index', compact('comments'));
     }
 }
